SuperFELTR with tests

+++ tests/individual_models/utils.py
@@ -10,36 +10,41 @@ import pickle
 
 @pytest.fixture(scope="session")
 def sample_dataset() -> Tuple[DrugResponseDataset, FeatureDataset, FeatureDataset]:
-    url = "https://zenodo.org/doi/10.5281/zenodo.12633909"
-    # Fetch the latest record
-    response = requests.get(url)
-    latest_url = response.links["linkset"]["url"]
-    response = requests.get(latest_url)
-    data = response.json()
-    name_to_url = {file["key"]: file["links"]["self"] for file in data["files"]}
-    tmpdir = tempfile.TemporaryDirectory()
-    toy_data_url = name_to_url["Toy_Data.zip"]
-    response = requests.get(toy_data_url)
-    file_path = os.path.join(tmpdir.name, "Toy_Data.zip")
+    # if the path ../../data/Toy_Data does not exist, download the data from Zenodo
+    if not os.path.exists("../data/Toy_Data"):
+        url = "https://zenodo.org/doi/10.5281/zenodo.12633909"
+        # Fetch the latest record
+        response = requests.get(url)
+        latest_url = response.links["linkset"]["url"]
+        response = requests.get(latest_url)
+        data = response.json()
+        name_to_url = {file["key"]: file["links"]["self"] for file in data["files"]}
+        tmpdir = tempfile.TemporaryDirectory()
+        dir_name = tmpdir.name
+        toy_data_url = name_to_url["Toy_Data.zip"]
+        response = requests.get(toy_data_url)
+        file_path = os.path.join(dir_name, "Toy_Data.zip")
 
-    print(f"Loading Toy Dataset from Zenodo, from {data['created']}")
+        print(f"Loading Toy Dataset from Zenodo, from {data['created']}")
 
-    with open(file_path, "wb") as f:
-        f.write(response.content)
+        with open(file_path, "wb") as f:
+            f.write(response.content)
 
-    with zipfile.ZipFile(file_path, "r") as zip_ref:
-        zip_ref.extractall(tmpdir.name)
+        with zipfile.ZipFile(file_path, "r") as zip_ref:
+            zip_ref.extractall(dir_name)
+    else:
+        dir_name = "../data"
 
     with open(
-        os.path.join(tmpdir.name, "Toy_Data", "toy_data_drp_dataset.pkl"), "rb"
+        os.path.join(dir_name, "Toy_Data", "toy_data_drp_dataset.pkl"), "rb"
     ) as f:
         drug_response = pickle.load(f)
     with open(
-        os.path.join(tmpdir.name, "Toy_Data", "toy_data_cl_features.pkl"), "rb"
+        os.path.join(dir_name, "Toy_Data", "toy_data_cl_features.pkl"), "rb"
     ) as f:
         cell_line_features = pickle.load(f)
     with open(
-        os.path.join(tmpdir.name, "Toy_Data", "toy_data_drug_features.pkl"), "rb"
+        os.path.join(dir_name, "Toy_Data", "toy_data_drug_features.pkl"), "rb"
     ) as f:
         drug_features = pickle.load(f)
     return drug_response, cell_line_features, drug_features
